separate add user from add manager files
