+++ .env.development.local
@@ -12,7 +12,7 @@ REACT_APP_URL_FE_ASSISTANCE=http://selfcare/assistance
 REACT_APP_URL_API_DASHBOARD=https://api.dev.selfcare.pagopa.it/dashboard
 REACT_APP_URL_API_ONBOARDING_V2=https://api.dev.selfcare.pagopa.it/onboarding
 
-REACT_APP_API_MOCK_PARTY_USERS=false
+REACT_APP_API_MOCK_PARTY_USERS=true
 REACT_APP_API_MOCK_PARTY_GROUPS=true
 
 REACT_APP_MAX_ADMIN_COUNT=4
