+++ src/pages/addUserFlow/addManager/__tests__/AddLegalRepresentativeForm.test.tsx
@@ -1,8 +1,8 @@
 
 import { fireEvent, screen } from '@testing-library/react';
-import { RoleEnum } from '../../../api/generated/onboarding/UserDto';
-import { mockedParties } from '../../../microcomponents/mock_dashboard/data/party';
-import { renderWithProviders } from '../../../utils/test-utils';
+import { RoleEnum } from '../../../../api/generated/onboarding/UserDto';
+import { mockedParties } from '../../../../microcomponents/mock_dashboard/data/party';
+import { renderWithProviders } from '../../../../utils/test-utils';
 import AddLegalRepresentativeForm from '../components/AddLegalRepresentativeForm';
 
 test('should render AddLegalRepresentativeForm with empty props', () => {
