+++ src/pages/addUserFlow/addUser/AddProductToUserPage.tsx
@@ -3,14 +3,14 @@ import { TitleBox } from '@pagopa/selfcare-common-frontend/lib';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
-import ProductNavigationBar from '../../components/ProductNavigationBar';
-import withUserDetail, { withUserDetailProps } from '../../decorators/withUserDetail';
-import { Party } from '../../model/Party';
-import { AddedUsersList, PartyUserOnCreation } from '../../model/PartyUser';
-import { Product } from '../../model/Product';
-import { ProductsRolesMap } from '../../model/ProductRole';
-import { RequestOutcomeMessage } from '../../model/UserRegistry';
-import AddLegalRepresentativeForm from './components/AddLegalRepresentativeForm';
+import ProductNavigationBar from '../../../components/ProductNavigationBar';
+import withUserDetail, { withUserDetailProps } from '../../../decorators/withUserDetail';
+import { Party } from '../../../model/Party';
+import { AddedUsersList, PartyUserOnCreation } from '../../../model/PartyUser';
+import { Product } from '../../../model/Product';
+import { ProductsRolesMap } from '../../../model/ProductRole';
+import { RequestOutcomeMessage } from '../../../model/UserRegistry';
+import AddLegalRepresentativeForm from '../addManager/AddLegalRepresentativeForm';
 import AddUserForm from './components/AddUserForm';
 import { MessageNoAction } from './components/MessageNoAction';
