+++ src/pages/addUserFlow/addUser/components/MessageNoAction.tsx
@@ -1,7 +1,7 @@
 import { Box, Grid, Typography, useTheme } from '@mui/material';
 import { useUnloadEventInterceptor } from '@pagopa/selfcare-common-frontend/lib/hooks/useUnloadEventInterceptor';
 import React from 'react';
-import { RequestOutcomeMessage } from '../../../model/UserRegistry';
+import { RequestOutcomeMessage } from '../../../../model/UserRegistry';
 
 export function MessageNoAction({ img, title, description, ImgComponent }: RequestOutcomeMessage) {
   const theme = useTheme();
